+++ xifire/src/main/java/com/xebia/xifire/model/SalarySoap.java
@@ -12,8 +12,8 @@ import java.util.List;
  * @generated
  */
 public class SalarySoap implements Serializable {
-    private int _id;
-    private int _userId;
+    private long _id;
+    private long _userId;
     private int _salaryMonth;
     private int _salaryYear;
     private int _basic;
@@ -97,27 +97,27 @@ public class SalarySoap implements Serializable {
         return soapModels.toArray(new SalarySoap[soapModels.size()]);
     }
 
-    public int getPrimaryKey() {
+    public long getPrimaryKey() {
         return _id;
     }
 
-    public void setPrimaryKey(int pk) {
+    public void setPrimaryKey(long pk) {
         setId(pk);
     }
 
-    public int getId() {
+    public long getId() {
         return _id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         _id = id;
     }
 
-    public int getUserId() {
+    public long getUserId() {
         return _userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(long userId) {
         _userId = userId;
     }
